requirements.txt
torch~=2.10.0+cu128
kornia~=0.7.1
codecarbon~=2.3.4
opencv-python~=4.9.0.80
numpy~=1.24.2
torchvision~=0.25.0+cu128
darmo~=0.1.15
timm~=0.9.16
pillow~=11.0.0
ptflops~=0.7.2.2
scikit-image~=0.22.0
scipy~=1.14.1
pandas~=2.2.3
tqdm~=4.67.0
matplotlib~=3.8.3
requests~=2.32.0
h5py~=3.10.0
gdown~=5.1.0